Fix math for counting beats ; add icons to chart editor
funny story, today is the first time I looked at the calculated 265 measures for descent and said "hm. at 134 bpm? that's a lot of notes. that would be like 4 or so minutes wouldn't it?" and realized my math was wrong. I was multiplying by 16 instead of 4 so take this as a lesson kids. understand your math